+++ src/main/scala/io/lunes/LunesNode.scala
@@ -19,7 +19,7 @@ import io.lunes.root.RootActorSystem
 import io.lunes.core.storage.db.openDB
 import io.lunes.features.api.ActivationApiRoute
 import io.lunes.core.storage.history.{CheckpointServiceImpl, StorageFactory}
-import io.lunes.http.NodeApiRoute
+import io.lunes.core.storage.http.NodeApiRoute
 import io.lunes.metrics.Metrics
 import io.lunes.mining.{Miner, MinerImpl}
 import io.lunes.network.RxExtensionLoader.RxExtensionLoaderShutdownHook
